+++ examples/tutorial/constant-folding/test.py
@@ -1,5 +1,0 @@
-a = 4
-print("hello")
-b = 9
-print(9)
-print("done")